+++ Logging/RockLib.Logging.Analyzers.Test/Verifiers/CSharpCodeFixVerifier`2.cs
@@ -34,7 +34,7 @@ namespace RockLib.Logging.Analyzers.Test
                 TestCode = source,
                 ReferenceAssemblies = ReferenceAssemblies.Default
                     .AddPackages(ImmutableArray.Create(
-                        new PackageIdentity("RockLib.Logging", "3.0.2")))
+                        new PackageIdentity("RockLib.Logging", "3.0.5")))
             };
 
             test.ExpectedDiagnostics.AddRange(expected);
@@ -58,7 +58,7 @@ namespace RockLib.Logging.Analyzers.Test
                 FixedCode = fixedSource,
                 ReferenceAssemblies = ReferenceAssemblies.Default
                     .AddPackages(ImmutableArray.Create(
-                        new PackageIdentity("RockLib.Logging", "3.0.2")))
+                        new PackageIdentity("RockLib.Logging", "3.0.5")))
             };
 
             test.ExpectedDiagnostics.AddRange(expected);