+++ Logging/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+#### Changed
+
+- RockLib0000: Include base classes when identifying public properties.
+- RockLib0000: Look for calls to `SafeToLogAttribute.Decorate` and `NotSafeToLogAttribute.Decorate` when determining if a type or property has been marked as safe to log.
+
 ## 1.0.1
 
 #### Added